requirement.txt
absl-py==2.1.0
apache-beam==2.61.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==2.2.1
coloredlogs==15.0.1
contourpy==1.3.1
crcmod==1.7
cycler==0.12.1
decorator==5.1.1
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
executing==2.1.0
fastavro==1.9.7
fasteners==0.19
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.5
h5py==3.12.1
hdfs==2.7.3
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipython==8.30.0
jedi==0.19.2
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
ml-dtypes==0.4.1
moviepy==2.1.1
mpmath==1.3.0
namex==0.0.8
natsort==8.4.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
objsize==0.7.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydot==1.4.2
Pygments==2.18.0
PyMatting==1.1.13
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.9.20
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
zstandard==0.23.0
